requirements.txt
monitorcontrol
pystray
Pillow
pynput
